+++ src/app/AppContainer.module.scss
@@ -1,0 +1,2 @@
+.app_container {
+}
