+++ src/app/page.module.scss
@@ -1,71 +1,3 @@
 .main {
-  $bubble-count: 15;
-  $sway-type: "sway-left-to-right", "sway-right-to-left";
   background-color: #f7f0e8;
-
-  @function random_range($min, $max) {
-    $rand: random();
-    $random_range: $min + floor($rand * (($max - $min) + 1));
-    @return $random_range;
-  }
-
-  @function sample($list) {
-    @return nth($list, random(length($list)));
-  }
-
-  .music_note_background {
-    position: relative;
-    width: 100%;
-    height: 100vh;
-    overflow: hidden;
-  }
-
-  .music_note {
-    position: absolute;
-    left: var(--bubble-left-offset);
-    bottom: -75%;
-    display: block;
-    width: var(--bubble-radius);
-    height: var(--bubble-radius);
-    animation: float-up var(--bubble-float-duration) var(--bubble-float-delay)
-      ease-in infinite;
-
-    @for $i from 0 through $bubble-count {
-      &:nth-child(#{$i}) {
-        --bubble-left-offset: #{random_range(0vw, 100vw)};
-        --bubble-radius: #{random_range(4vw, 8vw)};
-        --bubble-float-duration: #{random_range(6s, 12s)};
-        --bubble-sway-duration: #{random_range(4s, 6s)};
-        --bubble-float-delay: #{random_range(0s, 4s)};
-        --bubble-sway-delay: #{random_range(0s, 4s)};
-        --bubble-sway-type: #{sample($sway-type)};
-      }
-    }
-  }
-
-  @keyframes float-up {
-    to {
-      transform: translateY(-175vh);
-    }
-  }
-
-  @keyframes sway-left-to-right {
-    from {
-      transform: translateX(-100%);
-    }
-
-    to {
-      transform: translateX(100%);
-    }
-  }
-
-  @keyframes sway-right-to-left {
-    from {
-      transform: translateX(100%);
-    }
-
-    to {
-      transform: translateX(-100%);
-    }
-  }
 }
